+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Fixed
 
 - KWIC download sometimes fails because row.structs is optional [#492](https://github.com/spraakbanken/korp-frontend/issues/492)
+- Moved morphology labels in the lemgram autocomplete
 
 ## [9.12.0] – 2025-10-29