+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Fixed
 
+- Streamlined export controls [#154](https://github.com/spraakbanken/korp-frontend/issues/154)
 - KWIC download sometimes fails because row.structs is optional [#492](https://github.com/spraakbanken/korp-frontend/issues/492)
 - Moved morphology labels in the lemgram autocomplete
 - Sort attributes alphabetically in Extended search